modularizaod en parte
